+++ lib/candela_standard_app_profile.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class CandelaStandardAppProfile {
+  final IconData icon;
+  final double iconScale;
+  final String text;
+  final Color primary;
+  final Color secondary;
+  final bool invertIconGradient;
+
+  CandelaStandardAppProfile({
+    required this.icon,
+    required this.iconScale,
+    required this.text,
+    required this.primary,
+    required this.secondary,
+    required this.invertIconGradient,
+  });
+}