+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.2
+
+- Renamed to Candela Standard Window
+- Added Candela Standard Profile class
+  - Added prefabs for class
+- Added script for generating icons
+
 ## 0.0.1
 
 - Added Candela Standard App class.